Refactor Variable and Parameter structs
This commit renames the Variable and Parameter structs to VarDeclaration and ParamDeclaration which better match the grammar and the role of these objects. The commit also slightly refactor VarDeclaration to group the variable's type and its domain in a single Variable struct. That struct will be useful to add support for predicate parsing in a future commit.