Minor change to Expr + more godoc
Having the IsArray bool in Expr to specify whether the array should be treated as an array or a single item was confusing. I'm still not super happy with the current struct but I believe it does a better job at representing either a single item or a list of items. It is also more consistent with other types such as Literal.